Add files via upload
Simply a very long-winded footer I created with lots of text and three pictures to add to the end of my WP posts which I cross post to steem using SteemPress. The idea is to encourage more people to sign up to steem.... 

You should just be able to cut and past the 'code' below. In the upload file I've inlucded a Word doc containing the same 'code' and the pictures I created or borrowed. 

WordPress - Steem footer

<img class=" size-full wp-image-22297 alignright" src="https://revisesociology.com/wp-content/uploads/2018/11/Steemit-pictures.jpg" alt="Steemit pictures" width="269" height="275" /></a>This post is also being published to the <a href="https://steem.com/" target="_blank" rel="noopener">steem network</a>, which pays you in cryptocurrency (steem) for blogging, vlogging, commenting on and even just upvoting other people's content.

If you are a blogger, vlogger, entrepeneur or developer, or even if you just like good content, you should get on steem: it's social media 4.0!

Even though steem only went live in 2016, it has already evolved into a thriving and diverse, decentralised network, and there are a number of ways you can start earning from your content on steem:
<ul>
 	<li>You can blog on [steemit.com](https://signup.steemit.com/), which is like a cross between WordPress and Facebook</li>
 	<li>You can upload videos to [Dtube](https://d.tube/), which is like YouTube</li>
 	<li>You can upload pictures to [Appics](https://appics.com/), like instagram,</li>
</ul>
In addition to these well known social media formats, there are a number of other innovative applications being built on the steem network, and you can currently earn ‘steem’ by doing any or all of the following on the platform:
<ul>
 	<li>Going for runs and walks and posting proof of your steps (via @actifit)</li>
 	<li>Picking up litter in your local area and taking pictures or making videos of yourself doing so (via @cleanplanet)</li>
 	<li>Hunting cool products and posting brief reviews of them (via @steemhunt)</li>
 	<li>Playing a fantasy card game (via @steemmonsters, something like Magic the Gathering)</li>
 	<li>If you're into data analysis and software development, you can get rewarded for contributing to open source projects via [Utopian-io](https://join.utopian.io/)</li>
</ul>
<h3><a href="https://signup.steemit.com/"><img class="aligncenter size-full wp-image-22298" src="https://revisesociology.com/wp-content/uploads/2018/11/Screen-Shot-2018-11-22-at-11.32.30.png" alt="Screen Shot 2018-11-22 at 11.32.30" width="516" height="518" /></a></h3>
<h3></h3>
<h3><strong>Four reasons you should be using steem rather than mainstream social media</strong></h3>
<ol>
 	<li>There are a lot of very clever people on steem, all over the world. It is a truly global, truly connected network of people devoted to doing good things. The projects listed above are just the tip of the iceberg.</li>
 	<li>Steem is controlled by the people, rather than by a Corporation – all of the code, the rewards structure, the applications which run on it, are designed by ordinary (albeit very clever) people, rather than it being designed by Facebook engineers who create an environment conducive to maximising their own profit.</li>
 	<li>On steem, you actually get rewarded for your work - uploading, commenting and ‘liking’, rather than you getting nothing for interacting and that data being sold on by Facebook.</li>
 	<li>Steem is anti-censorship – unlike with Facebook, you will not have your content taken down because it might cause offense.</li>
</ol>
The steem network is barely 3 years old. It has the potential to be the next Facebook, Instagram, DTube, Amazon, Twitter and Udemy all on one blockchain. This means that on top of all of the reasons for joining given above, there is also huge potential for growth in the value of the steem cryptocurrency token.
<h3><strong>Sign up for steem</strong></h3>
To find out more about steem, you should visit steem's homepage - https://steem.com/ where you'll find a link to open up a new account.

Alternatively if you can't be fussed to click two links, then [this link](https://signup.steemit.com/) will take you straight to the steemit new users sign up page.

<a href="https://signup.steemit.com/"><img class=" size-full wp-image-22297 aligncenter" src="https://revisesociology.com/wp-content/uploads/2018/11/Steemit-pictures.jpg" alt="Steemit pictures" width="269" height="275" /></a>

FIN